# docs/api-coverage/openapi-non-prod.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
...
  "paths" : {
    "/api/agent/run" : {
      "post" : {
        "operationId" : "run_1",
        "parameters" : [ {
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AgentRunRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAgentRunVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "agent-controller" ]
      }
    },
    "/api/agent/runs" : {
      "get" : {
        "operationId" : "listRuns",
        "parameters" : [ {
          "in" : "query",
          "name" : "pageNum",
          "required" : false,
          "schema" : {
            "default" : 1,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "pageSize",
          "required" : false,
          "schema" : {
            "default" : 10,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "projectUuid",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "agentType",
          "required" : false,
          "schema" : {
            "enum" : [ "REQUIREMENT_BREAKDOWN", "API_DESIGN", "BUG_ANALYSIS", "PROMPT_GENERATE", "GAME_CONCEPT", "CORE_LOOP_DESIGN", "TASK_BREAKDOWN", "GAME_CONFIG_GENERATE", "BALANCE_EVALUATION" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "RUNNING", "SUCCESS", "FAILED", "TIMEOUT" ],
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageAgentRunVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "agent-controller" ]
      }
    },
    "/api/agent/runs/{runUuid}" : {
      "get" : {
        "operationId" : "getRun_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "runUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAgentRunVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "agent-controller" ]
      }
    },
    "/api/artifacts/{artifactUuid}" : {
      "get" : {
        "operationId" : "getArtifact",
        "parameters" : [ {
          "in" : "path",
          "name" : "artifactUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAgentArtifactVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "artifact-controller" ]
      }
    },
...
    "/api/projects" : {
      "get" : {
        "operationId" : "listProjects",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListGameProjectVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "game-controller" ]
      },
      "post" : {
        "operationId" : "createProject",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GameProjectRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseGameProjectVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "game-controller" ]
      }
    },
    "/api/projects/{projectUuid}" : {
      "get" : {
        "operationId" : "getProject",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseGameProjectVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "game-controller" ]
      },
      "put" : {
        "operationId" : "updateProject",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GameProjectRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseGameProjectVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "game-controller" ]
      }
    },
    "/api/projects/{projectUuid}/artifacts" : {
      "get" : {
        "operationId" : "listProjectArtifacts",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListAgentArtifactVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "artifact-controller" ]
      }
    },
    "/api/projects/{projectUuid}/artifacts/{artifactUuid}" : {
      "get" : {
        "operationId" : "getProjectArtifact",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "artifactUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAgentArtifactVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "artifact-controller" ]
      }
    },
...
    "/api/v1/workflow-runs/{workflowRunUuid}/artifacts" : {
      "get" : {
        "operationId" : "getArtifacts",
        "parameters" : [ {
          "in" : "path",
          "name" : "workflowRunUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListArtifactSummaryVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "workflow-run-query-controller" ]
      }
    },
...
    "/api/v5/projects/{projectUuid}/generation-runs/{runUuid}/artifact" : {
      "get" : {
        "operationId" : "artifact",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "runUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "format" : "byte",
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "generation-run-controller" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "AgentArtifactVO" : {
        "properties" : {
          "agentRunId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "artifactType" : {
            "type" : "string"
          },
          "artifactUuid" : {
            "type" : "string"
          },
          "content" : {
            "type" : "string"
          },
          "contentDigest" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "projectId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "runtimeCapabilityVersion" : {
            "type" : "string"
          },
          "runtimeEligible" : {
            "type" : "boolean"
          },
          "schemaKey" : {
            "type" : "string"
          },
          "schemaVersion" : {
            "type" : "string"
          },
          "sourceArtifactUuid" : {
            "type" : "string"
          },
          "sourceAttempt" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "validationSummary" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "AgentRunRequest" : {
        "properties" : {
          "agentType" : {
            "enum" : [ "REQUIREMENT_BREAKDOWN", "API_DESIGN", "BUG_ANALYSIS", "PROMPT_GENERATE", "GAME_CONCEPT", "CORE_LOOP_DESIGN", "TASK_BREAKDOWN", "GAME_CONFIG_GENERATE", "BALANCE_EVALUATION" ],
            "type" : "string"
          },
          "content" : {
            "minLength" : 1,
            "type" : "string"
          },
          "context" : {
            "maxLength" : 2000,
            "minLength" : 0,
            "type" : "string"
          },
          "projectUuid" : {
            "minLength" : 1,
            "type" : "string"
          },
          "ragContextBudget" : {
            "format" : "int32",
            "type" : "integer"
          },
          "ragEnabled" : {
            "type" : "boolean"
          },
          "ragTopK" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "agentType", "content", "projectUuid", "title" ],
        "type" : "object"
      },
...
      "AgentRunVO" : {
        "properties" : {
          "agentType" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "errorCategory" : {
            "type" : "string"
          },
          "errorMessage" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "mockState" : {
            "type" : "string"
          },
          "modelName" : {
            "type" : "string"
          },
          "projectId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "projectUuid" : {
            "type" : "string"
          },
          "provider" : {
            "type" : "string"
          },
          "runUuid" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "timeTakenMs" : {
            "format" : "int64",
            "type" : "integer"
          },
          "traceId" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "userId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "ApiResponseAgentArtifactVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/AgentArtifactVO"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiResponseAgentRunVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/AgentRunVO"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseGameProjectVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/GameProjectVO"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseListAgentArtifactVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "items" : {
              "$ref" : "#/components/schemas/AgentArtifactVO"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseListArtifactSummaryVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "items" : {
              "$ref" : "#/components/schemas/ArtifactSummaryVO"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiResponseListGameProjectVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "items" : {
              "$ref" : "#/components/schemas/GameProjectVO"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponsePageAgentRunVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/PageAgentRunVO"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ArtifactSummaryVO" : {
        "properties" : {
          "artifactUuid" : {
            "type" : "string"
          },
          "contentDigest" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "runtimeCapabilityVersion" : {
            "type" : "string"
          },
          "runtimeEligible" : {
            "type" : "boolean"
          },
          "schemaKey" : {
            "type" : "string"
          },
          "schemaVersion" : {
            "type" : "string"
          },
          "sourceArtifactUuid" : {
            "type" : "string"
          },
          "sourceAttempt" : {
            "format" : "int32",
            "type" : "integer"
          },
          "status" : {
            "type" : "string"
          },
          "stepKey" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          },
          "validationSummary" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "GameProjectRequest" : {
        "properties" : {
          "description" : {
            "minLength" : 1,
            "type" : "string"
          },
          "gameType" : {
            "minLength" : 1,
            "type" : "string"
          },
          "name" : {
            "minLength" : 1,
            "type" : "string"
          },
          "targetPlatform" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "description", "gameType", "name", "targetPlatform" ],
        "type" : "object"
      },
      "GameProjectVO" : {
        "properties" : {
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "gameType" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "projectUuid" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "targetPlatform" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "userId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "OrderItem" : {
        "properties" : {
          "asc" : {
            "type" : "boolean"
          },
          "column" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "PageAgentRunVO" : {
        "properties" : {
          "countId" : {
            "type" : "string",
            "writeOnly" : true
          },
          "current" : {
            "format" : "int64",
            "type" : "integer"
          },
          "maxLimit" : {
            "format" : "int64",
            "type" : "integer",
            "writeOnly" : true
          },
          "optimizeCountSql" : {
            "$ref" : "#/components/schemas/PageAgentRunVO",
            "writeOnly" : true
          },
          "optimizeJoinOfCountSql" : {
            "type" : "boolean",
            "writeOnly" : true
          },
          "orders" : {
            "items" : {
              "$ref" : "#/components/schemas/OrderItem"
            },
            "type" : "array",
            "writeOnly" : true
          },
          "pages" : {
            "format" : "int64",
            "type" : "integer"
          },
          "records" : {
            "items" : {
              "$ref" : "#/components/schemas/AgentRunVO"
            },
            "type" : "array"
          },
          "searchCount" : {
            "$ref" : "#/components/schemas/PageAgentRunVO",
            "writeOnly" : true
          },
          "size" : {
            "format" : "int64",
            "type" : "integer"
          },
          "total" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    }
  }
}